requirements.txt
Django==2.2.4
django-model-utils==4.0.0